Modularize entities
