+++ openfisca_core/entities/role.py
@@ -1,0 +1,16 @@
+import textwrap
+
+
+class Role:
+
+    def __init__(self, description, entity):
+        self.entity = entity
+        self.key = description['key']
+        self.label = description.get('label')
+        self.plural = description.get('plural')
+        self.doc = textwrap.dedent(description.get('doc', ""))
+        self.max = description.get('max')
+        self.subroles = None
+
+    def __repr__(self):
+        return "Role({})".format(self.key)
